#2 - Refactor: move domain to the domain file
