+++ src/main/kotlin/com/gmaur/investment/robotadvisor/RebalanceRequest.kt
@@ -1,6 +1,6 @@
 package com.gmaur.investment.robotadvisor
 
+import com.gmaur.investment.robotadvisor.domain.AssetAllocation
 import com.gmaur.investment.robotadvisor.domain.Portfolio
-import com.gmaur.investment.robotadvisor.infrastructure.AssetAllocation
 
 data class RebalanceRequest(val ideal: AssetAllocation, val current: Portfolio)
